move feed to validator type instead of profile
