+++ api/lib/reducers/cosmosV0-reducers.js
@@ -341,8 +341,7 @@ function validatorProfileReducer(
       numberStakers,
       uptimePercentage: validator.uptimePercentage,
       contributionLinks: JSON.parse(validatorProfile.contributionLinks),
-      network,
-      feed: validator.feed
+      network
     }
   }
 }
